changes in creating tables
