+++ backend/tests/DatabaseCreationTest.py
@@ -3,7 +3,7 @@ import pytest
 import sys
 import os
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-from DatabaseCreator import DatabaseCreator
+from backend.utils.DatabaseCreator import DatabaseCreator
 
 @pytest.fixture
 def db():
